Move react-native-editor hooks to dedicated hooks directory
